Add page to root.
The page is split into two sections, the header and the content. I use
a mobile-first approach; the two sections' positions change accoring to
screensize. That is, the mobile version of the page has the header on
top and the content on the bottom. However, on larger screens, the
header is on the left and the content is on the right.

The header consists of my name as the heading and my title as the
subheading. Below the header texts is the table of contents: (1) About,
(2) Projects, and (3) Contact. Clicking these links triggers scrolls to
their respective sections of content, while indicating the currently
active section. At the bottom of the header section, there is container
of all the languages and technologies I am comfortable using and have
used in prior projects.

All containers are flexbox containers, which aids in adding
responsiveness to my website.