Feat: Responsive active section indicator
As the end user scrolls, the currently shown section (e.g. about,
projects, contact) on screen. The heights of the main content sections
change as the user resizes the window. Therefore, the heights of each of
the sections need to be tracked/updated upon window resize; if not, the
indicators will not trigger at the correct thresholds when resized. So,
we keep track of when the user resizes the window and trigger a function
to update the heights. Using these new heights, we can correctly
calculate the split points. Prior, assumed the heights of each section
would be the same; now, heights can be different and the indicator will
reigster correctly in the navigation bar.